+++ src/id3/v2/items/mod.rs
@@ -1,2 +1,5 @@
 pub(super) mod encapsulated_object;
+pub(super) mod encoded_text_frame;
+pub(super) mod language_frame;
+pub(super) mod popularimeter;
 pub(super) mod sync_text;